pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Core_Analytics"
version = "0.1.0"
description = "Core_Analytics"
requires-python = ">=3.13"
dependencies = [
    "Jinja2==3.1.6",
    "apscheduler==3.11.0",
    "azure-identity==1.24.0",
    "azure-monitor-query==2.0.0",
    "azure-storage-blob==12.26.0",
    "fastapi==0.116.1",
    "gunicorn==23.0.0",
    "openpyxl==3.1.5",
    "pandas==2.3.2",
    "pyyaml==6.0.2",
    "uvicorn==0.35.0",
    "xlsxwriter==3.2.5",
    "sendgrid==6.12.5",
]

[dependency-groups]
dev = [
    "python-dotenv==1.1.1",
]


[tool.hatch.build.targets.wheel]
packages = ["."]
